Add functionality to update a vessel
Created route and controller with validation to update a record on the Vessel model. Enhanced validation to check when updating a vessel that its name is unique compared against other records. Added comments as explanation of the process for updating a record based on id from input rather than route parameter.