Add enhanced update port functionality
Modified UpdateVoyagePort controller with logic to prevent a user from updating a port if the port is currently being used as part of an active voyage. Enabled force delete to bypass this if required. Created associated tests for this functionality. Modified update port route to accept port id in URL instead of form request. Updated all tests affected by this change.