Update voyages/create endpoint
Updated migration and added guards to controller logic for creating voyage data: checking whether the title is a duplicate, the user/company is authorized to use the requested vessel and ports, and that voyage dates make sense chronologically and do not conflict with another voyage for the same vessel. Created associated tests.